Implement responsive menu
Menu now works in mobile and normal modes.
I added a second nav section for this.  With some work, we can probably get it back down to one,
but this should be fine until we can rewrite the site using React so we can avoid all the cut/paste repetition
between pages, and within pages.