Fix another minor NaN bug
Add server error string to the display when performing final submission to the /adduser endpoint (some users are getting a generic error here - this will help troubleshoot the issue).
Recalculate coupon discounts (incl. which coupons are or are not applied) when leaving and returning to the coupon screen (in case user adds/removes units from the order).
Refactor code so that it's easier to copy DCR1 app to SC1 (they now literally differ by just two lines at the top of the file.